requirements.txt
streamlit==1.28.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.35.0
torch==2.1.0
accelerate==0.24.0
PyPDF2==3.0.1
nltk==3.8.1
numpy==1.24.3
pandas==2.0.3
jupyter==1.0.0